requirements.txt
# ===== CORE RASA =====
rasa==3.6.21
rasa-sdk==3.6.2

# ===== ML CORE =====
tensorflow==2.12.0
numpy==1.23.5
protobuf==4.23.3
pydantic==1.10.9
packaging==20.9

# ===== RAG STACK =====
chromadb==0.4.24

# ===== EMBEDDING =====
sentence-transformers==2.6.1
transformers==4.35.2
torch==2.0.1
google-generativeai==0.3.2

# ==== STREAMLIT UI ====
streamlit==1.56.0

# ===== OPTIONAL =====
scikit-learn==1.1.3
scipy==1.10.1